OTH : Update tests
